+++ packages/browser/src/login/oidc/oidcHandlers/AuthorizationCodeWithPkceOidcHandler.spec.ts
@@ -162,7 +162,7 @@ describe("AuthorizationCodeWithPkceOidcHandler", () => {
         mockedStorage.getForUser("mySession", "redirectUrl", {
           secure: false,
         })
-      ).resolves.toStrictEqual("https://app.example.com?someQuery=someValue");
+      ).resolves.toBe("https://app.example.com?someQuery=someValue");
       await expect(
         mockedStorage.getForUser("mySession", "codeVerifier", {
           secure: false,
@@ -209,7 +209,7 @@ describe("AuthorizationCodeWithPkceOidcHandler", () => {
       );
     });
 
-    it("requests both openid and offline_access scopes", async () => {
+    it("requests both openid, offline_access and webid scopes", async () => {
       const oidcModule = mockOidcModule();
       const authorizationCodeWithPkceOidcHandler =
         getAuthorizationCodeWithPkceOidcHandler();
@@ -223,7 +223,7 @@ describe("AuthorizationCodeWithPkceOidcHandler", () => {
       await authorizationCodeWithPkceOidcHandler.handle(oidcOptions);
       expect(oidcModule.OidcClient).toHaveBeenCalledWith(
         expect.objectContaining({
-          scope: "openid offline_access",
+          scope: "openid offline_access webid",
         })
       );
     });