+++ packages/node/src/login/oidc/ClientRegistrar.spec.ts
@@ -99,10 +99,35 @@ describe("ClientRegistrar", () => {
           ...IssuerConfigFetcherFetchConfigResponse,
         }
       );
-      expect(client.clientId).toEqual("an id");
-      expect(client.clientSecret).toEqual("a secret");
-      expect(client.clientName).toEqual("my client name");
-      expect(client.idTokenSignedResponseAlg).toEqual("ES256");
+      expect(client.clientId).toBe("an id");
+      expect(client.clientSecret).toBe("a secret");
+      expect(client.clientName).toBe("my client name");
+      expect(client.idTokenSignedResponseAlg).toBe("ES256");
+    });
+
+    it("negotiates signing alg if not found in storage", async () => {
+      const clientRegistrar = getClientRegistrar({
+        storage: mockStorageUtility(
+          {
+            "solidClientAuthenticationUser:mySession": {
+              clientId: "an id",
+              clientSecret: "a secret",
+              clientName: "my client name",
+            },
+          },
+          false
+        ),
+      });
+      const client = await clientRegistrar.getClient(
+        {
+          sessionId: "mySession",
+          redirectUrl: "https://example.com",
+        },
+        {
+          ...IssuerConfigFetcherFetchConfigResponse,
+        }
+      );
+      expect(client.idTokenSignedResponseAlg).toBe("ES256");
     });
 
     it("properly performs dynamic registration and saves client information", async () => {