+++ packages/node/src/login/oidc/redirectHandler/FallbackRedirectHandler.spec.ts
@@ -55,7 +55,7 @@ describe("FallbackRedirectHandler", () => {
     it("returns an unauthenticated session", async () => {
       const redirectHandler = new FallbackRedirectHandler();
       const mySession = await redirectHandler.handle("https://my.app");
-      expect(mySession.isLoggedIn).toEqual(false);
+      expect(mySession.isLoggedIn).toBe(false);
       expect(mySession.webId).toBeUndefined();
     });
   });