+++ packages/node/src/login/oidc/refresh/TokenRefresher.spec.ts
@@ -24,16 +24,12 @@ import {
   mockStorageUtility,
   StorageUtilityMock,
   EVENTS,
+  KeyPair,
 } from "@inrupt/solid-client-authn-core";
-// Until there is a broader support for submodules exports in the ecosystem,
-// (e.g. jest supports them), we'll depend on an intermediary package that exports
-// a single ES module. The submodule exports should be kept commented out to make
-// it easier to transition back when possible.
-// import { JWK } from "jose/types";
-// import { parseJwk } from "jose/jwk/parse";
-import { JWK, parseJwk } from "@inrupt/jose-legacy-modules";
+import { JWK, importJWK } from "jose";
 import { IdTokenClaims, TokenSet } from "openid-client";
 import { EventEmitter } from "events";
+import { KeyObject } from "crypto";
 import TokenRefresher from "./TokenRefresher";
 import {
   mockClientRegistrar,
@@ -43,8 +39,10 @@ import {
   mockDefaultIssuerConfig,
   mockIssuerConfigFetcher,
 } from "../__mocks__/IssuerConfigFetcher";
+import { negotiateClientSigningAlg } from "../ClientRegistrar";
 
 jest.mock("openid-client");
+jest.mock("../ClientRegistrar");
 
 const mockJwk = (): JWK => {
   return {
@@ -58,9 +56,9 @@ const mockJwk = (): JWK => {
   };
 };
 
-const mockKeyPair = async () => {
+const mockKeyPair = async (): Promise<KeyPair> => {
   return {
-    privateKey: await parseJwk(mockJwk()),
+    privateKey: (await importJWK(mockJwk())) as KeyObject,
     // Use the same JWK for public and private key out of convenience, don't do
     // this in real life.
     publicKey: mockJwk(),
@@ -119,22 +117,17 @@ const mockDpopTokens = (): TokenSet => {
 const setupOidcClientMock = (tokenSet: TokenSet) => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const { Issuer } = jest.requireMock("openid-client") as any;
-  function clientConstructor() {
-    // this is untyped, which makes TS complain
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    this.refresh = jest.fn().mockResolvedValueOnce(tokenSet);
-  }
   const mockedIssuer = {
     metadata: mockDefaultIssuerConfig(),
-    Client: clientConstructor,
+    Client: jest.fn().mockReturnValue({
+      refresh: jest.fn().mockResolvedValueOnce(tokenSet as never),
+    }),
   };
   Issuer.mockReturnValueOnce(mockedIssuer);
+  return mockedIssuer;
 };
 
-const setupDefaultOidcClientMock = () => {
-  setupOidcClientMock(mockDpopTokens());
-};
+const setupDefaultOidcClientMock = () => setupOidcClientMock(mockDpopTokens());
 
 const mockDefaultStorageContent = {
   "solidClientAuthenticationUser:mySession": {
@@ -206,6 +199,13 @@ describe("TokenRefresher", () => {
 
   it("throws if a refresh token isn't provided", async () => {
     setupDefaultOidcClientMock();
+    const mockedModule = jest.requireMock("../ClientRegistrar") as {
+      negotiateClientSigningAlg: typeof negotiateClientSigningAlg;
+    };
+    mockedModule.negotiateClientSigningAlg = jest
+      .fn(negotiateClientSigningAlg)
+      .mockReturnValue("ES256");
+
     const mockedStorage = mockRefresherDefaultStorageUtility();
 
     const refresher = getTokenRefresher({
@@ -219,6 +219,12 @@ describe("TokenRefresher", () => {
 
   it("throws if a DPoP token is expected, but no DPoP key is provided", async () => {
     setupDefaultOidcClientMock();
+    const mockedModule = jest.requireMock("../ClientRegistrar") as {
+      negotiateClientSigningAlg: typeof negotiateClientSigningAlg;
+    };
+    mockedModule.negotiateClientSigningAlg = jest
+      .fn(negotiateClientSigningAlg)
+      .mockReturnValue("ES256");
     const mockedStorage = mockRefresherDefaultStorageUtility();
 
     const refresher = getTokenRefresher({
@@ -232,10 +238,12 @@ describe("TokenRefresher", () => {
     );
   });
 
-  // FIXME: this test brings coverage to 100%, but has no meaningful expect.
-  // For the time being, it is sufficient, but to be fixed soon.
   it("does not negotiate the signing algorithm if it is already set for the client", async () => {
     setupDefaultOidcClientMock();
+    const mockedModule = jest.requireMock("../ClientRegistrar") as {
+      negotiateClientSigningAlg: typeof negotiateClientSigningAlg;
+    };
+    mockedModule.negotiateClientSigningAlg = jest.fn();
     const refresher = getTokenRefresher({
       storageUtility: mockRefresherDefaultStorageUtility(),
       clientRegistrar: mockClientRegistrar({
@@ -246,13 +254,36 @@ describe("TokenRefresher", () => {
       }),
     });
 
-    const refreshedTokens = await refresher.refresh(
+    await refresher.refresh(
       "mySession",
       "some refresh token",
       await mockKeyPair()
     );
 
-    expect(refreshedTokens.accessToken).toBe(mockDpopTokens().access_token);
+    expect(mockedModule.negotiateClientSigningAlg).not.toHaveBeenCalled();
+  });
+
+  it("uses 'none' authentication if using Solid-OIDC client identifiers", async () => {
+    const mockedIssuer = setupDefaultOidcClientMock();
+    const refresher = getTokenRefresher({
+      clientRegistrar: mockClientRegistrar({
+        clientId: "https://some.client.identifier",
+        clientType: "solid-oidc",
+        idTokenSignedResponseAlg: "ES256",
+      }),
+    });
+
+    await refresher.refresh(
+      "mySession",
+      "some refresh token",
+      await mockKeyPair()
+    );
+
+    expect(mockedIssuer.Client).toHaveBeenCalledWith(
+      expect.objectContaining({
+        token_endpoint_auth_method: "none",
+      })
+    );
   });
 
   it("refreshes a DPoP token properly", async () => {
@@ -311,12 +342,12 @@ describe("TokenRefresher", () => {
       "some old refresh token",
       await mockKeyPair()
     );
-    expect(refreshedTokens.refreshToken).toEqual("some new refresh token");
+    expect(refreshedTokens.refreshToken).toBe("some new refresh token");
 
     // Check that the session information is stored in the provided storage
     await expect(
       mockedStorage.getForUser("mySession", "refreshToken")
-    ).resolves.toEqual("some new refresh token");
+    ).resolves.toBe("some new refresh token");
   });
 
   it("calls the refresh token rotation handler if one is provided", async () => {
@@ -338,7 +369,7 @@ describe("TokenRefresher", () => {
       mockEmitter
     );
 
-    expect(refreshedTokens.refreshToken).toEqual("some new refresh token");
+    expect(refreshedTokens.refreshToken).toBe("some new refresh token");
     expect(mockEmit).toHaveBeenCalledWith(
       EVENTS.NEW_REFRESH_TOKEN,
       "some new refresh token"