+++ packages/node/src/login/oidc/refresh/TokenRefresher.ts
@@ -108,6 +108,9 @@ export default class TokenRefresher implements ITokenRefresher {
     const client = new issuer.Client({
       client_id: clientInfo.clientId,
       client_secret: clientInfo.clientSecret,
+      token_endpoint_auth_method: clientInfo.clientSecret
+        ? "client_secret_basic"
+        : "none",
       id_token_signed_response_alg: clientInfo.idTokenSignedResponseAlg,
     });